Authentication now working for Admin and User

+++ src/main/java/com/shopApplication/repositories/RoleRepository.java
@@ -1,13 +1,13 @@
 package com.shopApplication.repositories;
 
+import com.shopApplication.models.Role;
 import com.shopApplication.security.ERole;
-import com.shopApplication.security.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Optional<Role> findByName(ERole name);
 }
